add an icon
git-svn-id: http://svn.navi.cx/misc/trunk/fyre@6533 9a66fdd2-3ab9-0310-9df6-91dd0304fdbc

+++ src/animation.c
@@ -358,7 +358,7 @@ void animation_store_chunk(AnimChunkState *state,
 	else {
 	    g_log(G_LOG_DOMAIN, G_LOG_LEVEL_WARNING,
 		  "Duration chunk is incorrectly sized, %ld bytes instead of %ld",
-		  length, sizeof(gdouble));
+		  (long) length, (long) sizeof(gdouble));
 	}
 	break;
